requirements.txt
Flask>=3.1,<4.0
Flask-Login>=0.6,<1.0
Flask-Migrate>=4.0,<5.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-WTF>=1.2,<2.0
email-validator>=2.2,<3.0
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<4.0
pytest>=8.3,<9.0
python-dotenv>=1.0,<2.0
